% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{lower_bounds}
\alias{lower_bounds}
\alias{upper_bounds}
\title{Retrieve the lower or upper bounds of a test function.}
\usage{
lower_bounds(fn)

upper_bounds(fn)
}
\arguments{
\item{fn}{[\code{\link{soo_function}}] Object of type \code{\link{soo_function}} to query.}
}
\value{
[\code{numeric}] Vector of lower or upper bounds of test function.
}
\description{
Retrieve the lower or upper bounds of a test function.
}
