\name{discus_function}
\alias{discus_function}
\title{Discus test function generator.}
\usage{
  discus_function(dimensions)
}
\arguments{
  \item{dimensions}{Size of parameter space.}
}
\value{
  A \code{soo_function}.
}
\description{
  The discus test function is similar to a high condition
  ellipsoid. It is defined as
}
\details{
  \deqn{f(x) = 10^6 x_1^2 + x'x.}
}

