% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/first_hitting_times.R
\name{first_hitting_times}
\alias{first_hitting_times}
\title{Return numerical vector of first hitting times, counted in
function evaluations.}
\usage{
first_hitting_times(x, target_levels)
}
\arguments{
\item{x}{Object from which the reached target values are to be
extracted. Methods for numeric vectors and
\code{record_target_values_function} class objects are provided.}

\item{target_levels}{Numerical vector of target levels which should
  ideally be reached.}
}
\description{
Return numerical vector of first hitting times, counted in
function evaluations.
}
\seealso{
\code{\link{recording_function}}
}

