% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_id.R
\name{function_id}
\alias{function_id}
\title{Function ID}
\usage{
function_id(fn)
}
\arguments{
\item{fn}{[\code{\link{soo_function}}] Function to name.}
}
\value{
ID of function. Guaranteed to be unique among all test
functions.
}
\description{
Get a short id for the function that can be used in filenames and
such. It is guaranteed that the ID contains only \dQuote{safe}
characters in the range {A-Z,a-z,0-9,_,-}.
}
