% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_name.R
\name{function_name}
\alias{function_name}
\title{Function name}
\usage{
function_name(fn)
}
\arguments{
\item{fn}{[\code{\link{soo_function}}] Function to name.}
}
\value{
Name of function.
}
\description{
Get a pretty function name for a benchmark function.
}
