\name{function_name}
\alias{function_name}
\alias{function_name.soo_function}
\title{Get a pretty function name for a benchmark function.}
\usage{
  function_name(fn)

  \method{function_name}{soo_function} (fn)
}
\arguments{
  \item{fn}{Function to name.}
}
\value{
  Name of function.
}
\description{
  Get a pretty function name for a benchmark function.
}

