% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_beale.R
\docType{data}
\name{generate_beale_function}
\alias{generate_beale_function}
\title{Generator for the Beale test function.}
\format{\preformatted{function ()  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "beale"
 - attr(*, "name")= chr "Beale test function"
}}
\usage{
generate_beale_function
}
\value{
A \code{soo_function}.
}
\description{
The function is defined as:
}
\details{
\deqn{
  f(x) =
   (1.5 - x_1 * (1 - x_2))^2
   + (2.25 - x_1 * (1 - x_2^2))^2
   + (2.625 - x_1 * (1-x_2^3))^2
}
}
\examples{
f <- generate_beale_function()
plot(f, rank=TRUE)
}
\keyword{datasets}

