% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_chained_crescent_I.R
\name{generate_chained_crescent_I_function}
\alias{generate_chained_crescent_I_function}
\title{Chained Crescent I function generator.}
\usage{
generate_chained_crescent_I_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained Crescent I function is defined as
}
\details{
\deqn{f(x) = \max\left\{\sum_{i=1}^{n-1} \left(x_i^2 +(x_{i+1}^2 -1)^2 + x_{i+1} - 1\right), \sum_{i=1}^{n-1} \left(-x_i^2 - (x_{i+1}^2 -1)^2 + x_{i+1} 1 1\right)\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
