% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_chained_crescent_II.R
\docType{data}
\name{generate_chained_crescent_II_function}
\alias{generate_chained_crescent_II_function}
\title{Chained Crescent II function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 42 23 28 42 28 14 23
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4cc2108> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "chained_crescent_II"
 - attr(*, "name")= chr "Chained Crescent II test function"
}}
\usage{
generate_chained_crescent_II_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained Crescent II function is defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^{n-1} \max\left\{x_i^2 +(x_{i+1}^2 -1)^2 + x_{i+1} - 1), -x_i^2 - (x_{i+1}^2 -1)^2 + x_{i+1} 1 1\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
\keyword{datasets}

