% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_chained_mifflin_I.R
\docType{data}
\name{generate_chained_mifflin_function}
\alias{generate_chained_mifflin_function}
\title{Chained Mifflin 2 function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "chained_mifflin_2"
 - attr(*, "name")= chr "Chained Mifflin 2 test function"
}}
\usage{
generate_chained_mifflin_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained Mifflin function 2 is defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^{n-1} \left(-x_i + 2(x_i^2 + x_{i+1}^2 - 1) + 1.75|x_i^2 + x_{i+1}^2 - 1|\right)}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
\keyword{datasets}

