% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_double_sum.R
\docType{data}
\name{generate_double_sum_function}
\alias{generate_double_sum_function}
\title{Double sum test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "double-sum"
 - attr(*, "name")= chr "Double sum test function"
}}
\usage{
generate_double_sum_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = \sum_{i=1}^n \left( \sum_{j=1}^{i} x_j \right)^2}
}
\references{
H.-P. Schwefel. Evolution and Optimum Seeking.
John Wiley & Sons, New York, 1995.
}
\keyword{datasets}

