% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_generalized_MAXQ.R
\name{generate_generalized_maxq_function}
\alias{generate_generalized_maxq_function}
\title{Generalized MAXQ test function generator.}
\usage{
generate_generalized_maxq_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The generalized MAXW test function is defined as
}
\details{
\deqn{f(x) = \max_{1 \leq i \neq n} x_i^2}
}
\references{
Haarala, M. and Miettinen, K. and Maelelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}

