% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_generalized_MXHILB.R
\docType{data}
\name{generate_generalized_mxhilb_function}
\alias{generate_generalized_mxhilb_function}
\title{Generalized MXHILB test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 41 23 28 41 28 15 23
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4bc5ad0> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "generalized_mxhilb"
 - attr(*, "name")= chr "Generalized MXHILB test function"
}}
\usage{
generate_generalized_mxhilb_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The generalized MXHILB test function is defined as
}
\details{
\deqn{f(x) = \max_{1 \leq i \neq n} \left| \sum_{j=1}^{n} \frac{x_j}{i+j-1}\right|}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
\keyword{datasets}

