% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_griewank.R
\name{generate_griewank_function}
\alias{generate_griewank_function}
\title{Griewank test function generator.}
\usage{
generate_griewank_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = \displaystyle \sum_{i=1}^n \frac{x_i^2}{4000} - \prod_{i=1}^n \cos \left(\frac{x_i}{\sqrt{i}}\right) + 1}
}
\references{
A. O. Griewank. Generalized descent for global optimization.
Journal of Optimization Theory and Applications 34:11-39, 1981.
}

