% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_himmelblau.R
\name{generate_himmelblau_function}
\alias{generate_himmelblau_function}
\title{Generator for the Himmelblau test function.}
\usage{
generate_himmelblau_function()
}
\value{
A \code{soo_function}.
}
\description{
The function is defined as:
}
\details{
\deqn{
  f(x) = (x_1^2 + x_2 - 11)^2 + (x_1 + x_2^2 - 7)^2
}
}
\examples{
f <- generate_himmelblau_function()
plot(f, rank=TRUE)

}
