% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_number_of_active_faces.R
\name{generate_number_of_active_faces_function}
\alias{generate_number_of_active_faces_function}
\title{Number of active faces test function generator.}
\usage{
generate_number_of_active_faces_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The number of active faces function test function is defined as
}
\details{
\deqn{f(x) = \max_{1 \leq i \leq n} \left\{\ln\left(\left|\sum_{i=1}^{n} x_i\right| + 1\right), \ln(|x_i| + 1)\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
