% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_rastrigin.R
\docType{data}
\name{generate_rastrigin_function}
\alias{generate_rastrigin_function}
\title{Rastrigin test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "rastrigin"
 - attr(*, "name")= chr "Rastrigin test function"
}}
\usage{
generate_rastrigin_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = \displaystyle 10n + \sum_{i=1}^n \left(x_i^2 - 10 \cos\left(2\pi x_i\right)\right)}
}
\references{
L. A. Rastrigin. Extremal control systems.
Theoretical Foundations of Engineering Cybernetics Series.
(in Russian), Nauka, Moscow, 1974.
}
\keyword{datasets}

