% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_rosenbrock.R
\docType{data}
\name{generate_rosenbrock_function}
\alias{generate_rosenbrock_function}
\title{Rosenbrock test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "rosenbrock"
 - attr(*, "name")= chr "Rosenbrock test function"
}}
\usage{
generate_rosenbrock_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space. Must be greater than 1.}
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = \displaystyle\sum_{i=1}^{n-1} \left(100\left(x_{i+1} - x_i^2\right)^2 + \left(x_i - 1\right)^2\right)}
}
\references{
H. H. Rosenbrock. An Automatic Method for Finding the
Greatest or Least Value of a Function.
The Computer Journal, 3(3):175-184, 1960.
}
\keyword{datasets}

