% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_sphere.R
\docType{data}
\name{generate_sphere_function}
\alias{generate_sphere_function}
\title{Sphere test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 18 29 27 1 29 1 18 27
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4774f10> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "sphere"
 - attr(*, "name")= chr "Sphere test function"
}}
\usage{
generate_sphere_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}]  Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The sphere function is arguably the simplest test function. It is
defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^n x_i^2}
}
\references{
K. D. De Jong. An analysis of the behavior of a class
of genetic adaptive systems. PhD thesis, Department of Computer
and  Communication Sciences, University of Michigan, Ann Arbor,
1975.
}
\keyword{datasets}

