% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_minimum.R
\name{global_minimum}
\alias{global_minimum}
\title{Global Optimum}
\usage{
global_minimum(fn)
}
\arguments{
\item{fn}{Function to query.}
}
\value{
List with two elements. \code{par} contains the location
of the global minimum in the parameter space (possibly as a list
if there are multiple global minima) and \code{value} the function
value of the global minimum.
}
\description{
Retrieve the global minimum of a test function.
}
