\name{goldstein_price_function}
\alias{goldstein_price_function}
\title{Generator for the Goldstein-Price function.}
\usage{
  goldstein_price_function()
}
\value{
  A \code{soo_function}.
}
\description{
  Generator for the Goldstein-Price function.
}
\examples{
f <- ackley_function(2)
plot(f, rank=TRUE)
}

