% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/counting_function.R
\name{is_counting_function}
\alias{is_counting_function}
\title{Counting function}
\usage{
is_counting_function(fn)
}
\arguments{
\item{fn}{[\code{function}] Function to check.}
}
\description{
Check if a function or one of its wrapped functions is a
\sQuote{counting_function}.
}

