% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recording_function.R
\name{is_recording_function}
\alias{is_recording_function}
\title{Recording function}
\usage{
is_recording_function(fn)
}
\arguments{
\item{fn}{Function to check.}
}
\value{
\code{TRUE} if \code{fn} is a \sQuote{recording_function}
else \code{FALSE}.
}
\description{
Check if a function supports the \sQuote{recording_function} interface.
}

