\name{is_soo_function}
\alias{is_soo_function}
\title{Check if a function is a SOO function.}
\usage{
  is_soo_function(fn)
}
\arguments{
  \item{fn}{Function to check.}
}
\value{
  \code{TRUE} if \code{fn} is a proper SOO function, else
  \code{FALSE}.
}
\description{
  Check if a function is a SOO function.
}

