% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_soo_function_generator.R
\name{is_soo_function_generator}
\alias{is_soo_function_generator}
\title{SOO function generator}
\usage{
is_soo_function_generator(fn)
}
\arguments{
\item{fn}{Function.}
}
\value{
\code{TRUE} if \code{fn} is a function generator and
  \code{FALSE} otherwise.
}
\description{
Returns wether \code{fn} is a generator for a single objective
optimization function.
}

