% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/number_of_parameters.R
\name{number_of_parameters}
\alias{number_of_parameters}
\title{Number of parameters}
\usage{
number_of_parameters(fn)
}
\arguments{
\item{fn}{[\code{soo_function}]\cr Function.}
}
\value{
Expected length of first argument. I.e. the size of the
parameter space of the function \code{fn}.
}
\description{
Return the parameter space size of a function.
}

