\name{number_of_parameters}
\alias{number_of_parameters}
\alias{number_of_parameters.soo_function}
\title{Return the parameter space size of a function.}
\usage{
  number_of_parameters(fn)

  \method{number_of_parameters}{soo_function} (fn)
}
\arguments{
  \item{fn}{Function.}
}
\value{
  Expected length of first argument. I.e. the size of the
  parameter space of the function \code{fn}.
}
\description{
  Return the parameter space size of a function.
}

