% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_soo_function.R
\name{print.soo_function}
\alias{print.soo_function}
\title{Print a SOO function.}
\usage{
\method{print}{soo_function}(x, ...)
}
\arguments{
\item{x}{[\code{\link{soo_function}}] A \code{\link{soo_function}} object.}

\item{...}{Ignored.}
}
\description{
Print a SOO function.
}
