% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recording_function.R
\name{recorded_values}
\alias{recorded_values}
\title{Return the internally recorded parameter settings and function
values for a test function.}
\usage{
recorded_values(fn)
}
\arguments{
\item{fn}{A function whos recorded parameter settings and function
values should be returnd.}
}
\value{
A list with elements \code{time}, \code{par} and
\code{value} which are the point in time (in number of function
evaluations), the parameter setting and the function value
recorded. The parameter settings are stored in a matrix where each
column is one parameter settings.
}
\description{
Return the internally recorded parameter settings and function
values for a test function.
}

