\name{sphere_function}
\alias{sphere_function}
\title{Sphere test function generator.}
\usage{
  sphere_function(dimensions)
}
\arguments{
  \item{dimensions}{Size of parameter space.}
}
\value{
  A \code{soo_function}.
}
\description{
  The sphere function is arguably the simplest test
  function. It is defined as
}
\details{
  \deqn{f(x) = x'x.}
}

