% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_beale.R
\name{generate_beale_function}
\alias{generate_beale_function}
\title{Generator for the Beale test function.}
\usage{
generate_beale_function()
}
\value{
A \code{soo_function}.
}
\description{
The function is defined as:
}
\details{
\deqn{
  f(x) =
   (1.5 - x_1 * (1 - x_2))^2
   + (2.25 - x_1 * (1 - x_2^2))^2
   + (2.625 - x_1 * (1-x_2^3))^2
}
}
\examples{
f <- generate_beale_function()
plot(f, rank=TRUE)

}
