% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_chained_CB3_I.R
\name{generate_chained_cb3_i_function}
\alias{generate_chained_cb3_i_function}
\title{Chained CB3 I test function generator.}
\usage{
generate_chained_cb3_i_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained CB3 I test function is defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^{n-1} \max\left\{x_i^4 + x_{i+1}^2, (2-x_i)^2 + (2-x_{i+1})^2, 2e^{-x_i + x_{i+1}}\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
