% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_chained_CB3_II.R
\docType{data}
\name{generate_chained_cb3_ii_function}
\alias{generate_chained_cb3_ii_function}
\title{Chained CB3 II test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "chained_cb3_ii"
 - attr(*, "name")= chr "Chained CB3 II test function"
}}
\usage{
generate_chained_cb3_ii_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained CB3 II test function is defined as
}
\details{
\deqn{f(x) = \max\left\{\sum_{i=1}^{n-1} \left(x_i^4 + x_{i+1}^2\right), \sum_{i=1}^{n-1} \left((2-x_i)^2 + (2-x_{i+1})^2\right), \sum_{i=1}^{n-1} \left(2e^{-x_i + x_{i+1}}\right)\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
\keyword{datasets}

