% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_chained_crescent_II.R
\name{generate_chained_crescent_II_function}
\alias{generate_chained_crescent_II_function}
\title{Chained Crescent II function generator.}
\usage{
generate_chained_crescent_II_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The chained Crescent II function is defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^{n-1} \max\left\{x_i^2 +(x_{i+1}^2 -1)^2 + x_{i+1} - 1), -x_i^2 - (x_{i+1}^2 -1)^2 + x_{i+1} 1 1\right\}}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}

