% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_generalized_MXHILB.R
\name{generate_generalized_mxhilb_function}
\alias{generate_generalized_mxhilb_function}
\title{Generalized MXHILB test function generator.}
\usage{
generate_generalized_mxhilb_function(dimensions)
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The generalized MXHILB test function is defined as
}
\details{
\deqn{f(x) = \max_{1 \leq i \neq n} \left| \sum_{j=1}^{n} \frac{x_j}{i+j-1}\right|}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}

