% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_happycat.R
\docType{data}
\name{generate_happycat_function}
\alias{generate_happycat_function}
\title{Generator for the Happycat function.}
\format{\preformatted{function (dimensions, alpha = 0.125)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 28 31 43 1 31 1 28 43
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4ca48f8> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "happycat"
 - attr(*, "name")= chr "HappyCat test function"
}}
\usage{
generate_happycat_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}

\item{alpha}{[\code{numeric(1)}] Parameter for control of groove shape.}
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{
  f(x) =
    \left((x^Tx - n)^2\right)^\alpha
    + \frac{1}{N}\left(
      \frac{1}{2}x^Tx + \sum_{i=1}^n x_i
    \right) + \frac{1}{2}}
}
\examples{
f <- generate_happycat_function(2, 1/8)
plot(f)
}
\references{
H.-G. Beyer and S. Finck. HappyCat - A Simple Function Class
  Where Well-Known Direct Search Algorithms Do Fail.
  In: PPSN XII (Parallel Problem Solving from Nature), 367-376,
  Springer, Heidelberg, 2012.
}
\keyword{datasets}

