% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_himmelblau.R
\docType{data}
\name{generate_himmelblau_function}
\alias{generate_himmelblau_function}
\title{Generator for the Himmelblau test function.}
\format{\preformatted{function ()  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 17 33 27 28 33 28 17 27
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4b702c8> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "himmelblau"
 - attr(*, "name")= chr "Himmelblau test function"
}}
\usage{
generate_himmelblau_function
}
\value{
A \code{soo_function}.
}
\description{
The function is defined as:
}
\details{
\deqn{
  f(x) = (x_1^2 + x_2 - 11)^2 + (x_1 + x_2^2 - 7)^2
}
}
\examples{
f <- generate_himmelblau_function()
plot(f, rank=TRUE)
}
\keyword{datasets}

