% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_kotanchek.r
\docType{data}
\name{generate_kotanchek_function}
\alias{generate_kotanchek_function}
\title{Kotanchek test function generator.}
\format{\preformatted{function ()  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 32 18 59 32 59 10 18
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x2819e88> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "kotanchek"
 - attr(*, "name")= chr "Kotanchek test function"
}}
\usage{
generate_kotanchek_function
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = - \frac{\exp\left(-(x_1 - 1)^2\right)}{1.2 + \left(x_1 - 2.5\right)^2}}
}
\keyword{datasets}

