% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_kotanchek.r
\name{generate_kotanchek_function}
\alias{generate_kotanchek_function}
\title{Kotanchek test function generator.}
\usage{
generate_kotanchek_function()
}
\value{
A \code{soo_function}.
}
\description{
The definition used is
}
\details{
\deqn{f(x) = - \frac{\exp\left(-(x_1 - 1)^2\right)}{1.2 + \left(x_1 - 2.5\right)^2}}
}

