% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_mexican_hat.R
\docType{data}
\name{generate_mexican_hat_function}
\alias{generate_mexican_hat_function}
\title{Mexican hat test function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 34 23 29 34 29 15 23
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x49e75d8> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "mexican-hat"
 - attr(*, "name")= chr "Mexican hat test function"
}}
\usage{
generate_mexican_hat_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The Mexican hat function is defined slightly differently by
different people. The definition used here is
}
\details{
\deqn{f(x) = -(1 - x'x) * \exp\left(-\frac{x'x}{2}\right)}{f(x) = -(1 - x'x) * exp(-0.5 * x'x)}

Note that we have flipped the sign of the function so that it is a
minimization problem like all other SOO functions.
}
\keyword{datasets}

