% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f_nonsmooth_brown_two.R
\docType{data}
\name{generate_nonsmooth_generalized_brown_2_function}
\alias{generate_nonsmooth_generalized_brown_2_function}
\title{Nonsmooth generalized Brown function 2 function generator.}
\format{\preformatted{function (dimensions)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 52 23 28 52 28 15 23
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x4984d08> 
 - attr(*, "class")= chr "soo_function_generator"
 - attr(*, "id")= chr "nonsmooth_generalized_brown_2"
 - attr(*, "name")= chr "Nonsmooth generalized Brown 2 test function"
}}
\usage{
generate_nonsmooth_generalized_brown_2_function
}
\arguments{
\item{dimensions}{[\code{integer(1)}] Size of parameter space.}
}
\value{
A \code{soo_function}.
}
\description{
The generalized Brown function 2 is defined as
}
\details{
\deqn{f(x) = \sum_{i=1}^{n-1} \left(|x_i|^{x_{i+1}^2+1} + |x_{i+1}|^{x_i^2 + 1}\right)}
}
\references{
Haarala, M. and Miettinen, K. and Maekelae, M. M., New limited
memory bundle method for large-scale nonsmooth optimization.
}
\keyword{datasets}

