% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soo_function.R
\name{is_soo_function}
\alias{is_soo_function}
\title{Single Objective Optimization Function}
\usage{
is_soo_function(fn)
}
\arguments{
\item{fn}{Function to check.}
}
\value{
\code{TRUE} if \code{fn} is a proper SOO function, else \code{FALSE}.
}
\description{
Check if a function is a SOO function.
}
