% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_function.R
\name{number_of_evaluations}
\alias{number_of_evaluations}
\title{Retrieve evaluation counter}
\usage{
number_of_evaluations(fn)
}
\arguments{
\item{fn}{[\code{counting_function}]\cr A counting function  as
returned by \code{\link{counting_function}}.}
}
\value{
The current value of the evaluation counter.
}
\description{
Return the number of times a test function has been evaluated.
}
\details{
The test function must be wrapped by
\code{\link{counting_function}} for this to work.
}
