% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soo_function.R
\name{plot.soo_function}
\alias{plot.soo_function}
\title{Plot a test function in 1 or 2 dimensions.}
\usage{
\method{plot}{soo_function}(x, ...)
}
\arguments{
\item{x}{[\code{\link{soo_function}}] Function to plot.}

\item{...}{Passed to the respective plot function.}
}
\description{
Plot a test function in 1 or 2 dimensions.
}
\examples{
par(mfrow=c(2, 2))
fn <- generate_sphere_function(2)
plot(fn)
plot(fn, show="contour")
plot(fn, rank=TRUE)
plot(fn, log=TRUE)
}
\author{
Olaf Mersmann \email{olafm@p-value.net}
}
