% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/counting_function.R
\name{reset_evaluation_counter}
\alias{reset_evaluation_counter}
\title{Reset evaluation counter}
\usage{
reset_evaluation_counter(fn)
}
\arguments{
\item{fn}{[\code{counting_function}]\cr A counting function as
returned by \code{\link{counting_function}}.}
}
\value{
The current value of the evaluation counter.
}
\description{
Reset the evaluation counter of a test function.
}
\details{
The test function must be wrapped by
\code{\link{counting_function}} for this to work.
}

